requirements.txt
aiodns==2.0.0
aiohttp==3.7.4.post0
aiohttp-scraper==0.1.4
aioredis==1.3.1
aiosqlite==0.16.1
async-timeout==3.0.1
asyncpg==0.22.0
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
hiredis==2.0.0
idna==2.10
iso8601==0.1.14
multidict==5.1.0
pycares==3.1.1
pycparser==2.20
pypika-tortoise==0.1.0
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
requests-file==1.5.1
six==1.15.0
tldextract==2.2.3
tortoise-orm==0.17.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.4
yarl==1.6.3
